add strain dependent rheology module
