+++ doc/modules/changes/20190912_naliboff
@@ -1,0 +1,4 @@
+New: There is a new model in the rheology class for material behavior
+that depends on finite strain. At present, the functionality in this
+model deals strictly with strain softening behavior, which was 
+transferred from the visco plastic material model.
